requirements_streamlit.txt
# Streamlit Dashboard Requirements
streamlit>=1.28.0
plotly>=5.15.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
psutil>=5.9.0

# Existing project requirements
tensorflow>=2.13.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pymongo>=4.4.0
gridfs>=0.1.0
